requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
bcrypt==3.2.0; python_version >= '3.6'
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
click==7.1.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
colormath==3.0.0
colorthief==0.2.1
cycler==0.10.0
decorator==4.4.2; python_version >= '3.5'
dlib==19.21.1
dnspython==1.16.0
flask-bcrypt==0.7.1
flask-cors==3.0.10
flask-jwt-extended==4.1.0
flask==1.1.2
idna==2.10; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
imageio==2.9.0
itsdangerous==1.1.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
jinja2==2.11.3; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
joblib==1.0.1; python_version >= '3.6'
kiwisolver==1.3.1; python_version >= '3.6'
markupsafe==1.1.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
matplotlib==3.4.1; python_version >= '3.7'
networkx==2.5.1; python_version >= '3.6'
numpy==1.20.2
opencv-python==4.5.1.48
pillow==8.1.0
pycparser==2.20; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pyjwt==2.0.1; python_version >= '3.6'
pymongo[srv]==3.11.3
pyparsing==2.4.7; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-dateutil==2.8.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-dotenv==0.17.0
pywavelets==1.1.1; python_version >= '3.5'
requests==2.25.1
scikit-image==0.18.1
scikit-learn==0.24.1; python_version >= '3.6'
scipy==1.6.2; python_version < '3.10' and python_version >= '3.7'
six==1.15.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sklearn==0.0
threadpoolctl==2.1.0; python_version >= '3.5'
tifffile==2021.4.8; python_version >= '3.7'
urllib3==1.26.4; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'
werkzeug==1.0.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
